requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asyncpg==0.27.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
loguru==0.6.0
numpy==1.24.1
Pillow==9.4.0
pycparser==2.21
pydantic==1.10.4
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.22.0
telegraph==2.2.0
typing_extensions==4.4.0
urllib3==1.26.14
win32-setctime==1.1.0
